feat: add column color customization
- Extend Column model with optional `color` field (client types, server schema, DB migration)
- Update API endpoints to include `color` in request/response payloads
- Add server-side validation for allowed palette names
- Implement UI color picker in ColumnSettings with palette preview, selection, and removal
- Apply selected color styles to column components via Tailwind class lookup
- Define CSS custom properties for new color palettes in `index.css`
- Update tests to cover new `color` field across client and server
- Adjust related type definitions and helper functions to propagate `color` data.

+++ server/src/routes/cards.ts
@@ -437,7 +437,7 @@ cardsRouter.post(
 			       ELSE started_at
 			     END,
 			     done_at = CASE WHEN $4 THEN COALESCE(done_at, now()) ELSE NULL END,
-			     assignee_id = CASE WHEN $6 AND $7 IS NOT NULL AND NOT $8 THEN $7 ELSE assignee_id END
+			 assignee_id = CASE WHEN $6 AND $7::integer IS NOT NULL AND NOT $8 THEN $7::integer ELSE assignee_id END
 			   WHERE id = $1`,
 				[
 					cardId,
